requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
fastparquet
apache-airflow-providers-postgres
psycopg2-binary
dbt-postgres